state: sync Wagmi state with Renegade state (#267)
