+++ providers/wagmi-provider/theme.ts
@@ -1,0 +1,14 @@
+export const connectKitTheme = {
+  "--ck-body-background": "hsl(var(--background))",
+  "--ck-border-radius": "0",
+  "--ck-font-family": "var(--font-sans-extended)",
+  "--ck-primary-button-background": "hsl(var(--background))",
+  "--ck-primary-button-border-radius": "0",
+  "--ck-body-color": "hsl(var(--foreground))",
+  "--ck-body-color-muted": "hsl(var(--muted-foreground))",
+  "--ck-body-color-muted-hover": "hsl(var(--foreground))",
+  "--ck-qr-dot-color": "hsl(var(--chart-blue))",
+  "--ck-secondary-button-background": "hsl(var(--background))",
+  "--ck-qr-border-color": "hsl(var(--border))",
+  "--ck-overlay-background": "rgba(0,0,0,.8)",
+}
